feat: chain extension
feat: chain extensions
