+++ node/Cargo.toml
@@ -45,6 +45,8 @@ sc-consensus = { workspace = true }
 sc-consensus-grandpa = { workspace = true }
 sc-consensus-grandpa-rpc = { workspace = true }
 sp-consensus-grandpa = { workspace = true }
+sc-chain-spec-derive = { workspace = true }
+sc-chain-spec = { workspace = true }
 sc-client-api = { workspace = true }
 sp-runtime = { workspace = true }
 sp-io = { workspace = true }
