+++ node/src/chain_spec/localnet.rs
@@ -1,0 +1,119 @@
+// Allowed since it's actually better to panic during chain setup when there is an error
+#![allow(clippy::unwrap_used)]
+
+use super::*;
+
+pub fn localnet_config() -> Result<ChainSpec, String> {
+    let wasm_binary = WASM_BINARY.ok_or_else(|| "Development wasm not available".to_string())?;
+
+    // Give front-ends necessary data to present to users
+    let mut properties = sc_service::Properties::new();
+    properties.insert("tokenSymbol".into(), "TAO".into());
+    properties.insert("tokenDecimals".into(), 9.into());
+    properties.insert("ss58Format".into(), 13116.into());
+
+    Ok(ChainSpec::builder(
+        wasm_binary,
+        Extensions {
+            bad_blocks: Some(HashSet::from_iter(vec![
+                // Example bad block
+                H256::from_str(
+                    "0xc174d485de4bc3813ac249fe078af605c74ff91d07b0a396cf75fa04f81fa312",
+                )
+                .unwrap(),
+            ])),
+            ..Default::default()
+        },
+    )
+    .with_name("Bittensor")
+    .with_id("bittensor")
+    .with_chain_type(ChainType::Development)
+    .with_genesis_config_patch(localnet_genesis(
+        // Initial PoA authorities (Validators)
+        // aura | grandpa
+        vec![
+            // Keys for debug
+            authority_keys_from_seed("Alice"),
+            authority_keys_from_seed("Bob"),
+        ],
+        // Pre-funded accounts
+        true,
+    ))
+    .with_properties(properties)
+    .build())
+}
+
+fn localnet_genesis(
+    initial_authorities: Vec<(AuraId, GrandpaId)>,
+    _enable_println: bool,
+) -> serde_json::Value {
+    let mut balances = vec![
+        (
+            get_account_id_from_seed::<sr25519::Public>("Alice"),
+            1000000000000000u128,
+        ),
+        (
+            get_account_id_from_seed::<sr25519::Public>("Bob"),
+            1000000000000000u128,
+        ),
+        (
+            get_account_id_from_seed::<sr25519::Public>("Charlie"),
+            1000000000000000u128,
+        ),
+        (
+            get_account_id_from_seed::<sr25519::Public>("Dave"),
+            2000000000000u128,
+        ),
+        (
+            get_account_id_from_seed::<sr25519::Public>("Eve"),
+            2000000000000u128,
+        ),
+        (
+            get_account_id_from_seed::<sr25519::Public>("Ferdie"),
+            2000000000000u128,
+        ),
+    ];
+
+    // Check if the environment variable is set
+    if let Ok(bt_wallet) = env::var("BT_DEFAULT_TOKEN_WALLET") {
+        if let Ok(decoded_wallet) = Ss58Codec::from_ss58check(&bt_wallet) {
+            balances.push((decoded_wallet, 1_000_000_000_000_000u128));
+        } else {
+            eprintln!("Invalid format for BT_DEFAULT_TOKEN_WALLET.");
+        }
+    }
+
+    let trimvirate_members: Vec<AccountId> = bounded_vec![
+        get_account_id_from_seed::<sr25519::Public>("Alice"),
+        get_account_id_from_seed::<sr25519::Public>("Bob"),
+        get_account_id_from_seed::<sr25519::Public>("Charlie"),
+    ];
+
+    let senate_members: Vec<AccountId> = bounded_vec![
+        get_account_id_from_seed::<sr25519::Public>("Dave"),
+        get_account_id_from_seed::<sr25519::Public>("Eve"),
+        get_account_id_from_seed::<sr25519::Public>("Ferdie"),
+    ];
+
+    serde_json::json!({
+        "balances": { "balances": balances },
+        "aura": {
+            "authorities": initial_authorities.iter().map(|x| (x.0.clone())).collect::<Vec<_>>()
+        },
+        "grandpa": {
+            "authorities": initial_authorities
+                .iter()
+                .map(|x| (x.1.clone(), 1))
+                .collect::<Vec<_>>()
+        },
+        "sudo": {
+            "key": Some(get_account_id_from_seed::<sr25519::Public>("Alice"))
+        },
+        "triumvirateMembers": {
+            "members": trimvirate_members
+        },
+        "senateMembers": {
+            "members": senate_members,
+        },
+    })
+}
