+++ node/src/chain_spec/mod.rs
@@ -1,0 +1,84 @@
+// Allowed since it's actually better to panic during chain setup when there is an error
+#![allow(clippy::unwrap_used)]
+
+pub mod finney;
+pub mod localnet;
+pub mod testnet;
+
+use node_subtensor_runtime::{AccountId, Block, RuntimeGenesisConfig, Signature, WASM_BINARY};
+use sc_chain_spec_derive::ChainSpecExtension;
+use sc_service::ChainType;
+use sp_consensus_aura::sr25519::AuthorityId as AuraId;
+use sp_consensus_grandpa::AuthorityId as GrandpaId;
+use sp_core::crypto::Ss58Codec;
+use sp_core::{bounded_vec, sr25519, Pair, Public, H256};
+use sp_runtime::traits::{IdentifyAccount, Verify};
+use sp_runtime::AccountId32;
+use std::collections::HashSet;
+use std::env;
+use std::str::FromStr;
+
+// The URL for the telemetry server.
+// const STAGING_TELEMETRY_URL: &str = "wss://telemetry.polkadot.io/submit/";
+
+/// Node `ChainSpec` extensions.
+#[derive(Default, Clone, Serialize, Deserialize, ChainSpecExtension)]
+#[serde(rename_all = "camelCase")]
+pub struct Extensions {
+    /// Block numbers with known hashes.
+    pub fork_blocks: sc_client_api::ForkBlocks<Block>,
+    /// Known bad block hashes.
+    pub bad_blocks: sc_client_api::BadBlocks<Block>,
+}
+
+/// Specialized `ChainSpec`. This is a specialization of the general Substrate ChainSpec type.
+pub type ChainSpec = sc_service::GenericChainSpec<RuntimeGenesisConfig, Extensions>;
+
+/// Generate a crypto pair from seed.
+pub fn get_from_seed<TPublic: Public>(seed: &str) -> <TPublic::Pair as Pair>::Public {
+    TPublic::Pair::from_string(&format!("//{}", seed), None)
+        .expect("static values are valid; qed")
+        .public()
+}
+
+type AccountPublic = <Signature as Verify>::Signer;
+
+/// Generate an account ID from seed.
+pub fn get_account_id_from_seed<TPublic: Public>(seed: &str) -> AccountId
+where
+    AccountPublic: From<<TPublic::Pair as Pair>::Public>,
+{
+    AccountPublic::from(get_from_seed::<TPublic>(seed)).into_account()
+}
+
+/// Generate an Aura authority key.
+pub fn authority_keys_from_seed(s: &str) -> (AuraId, GrandpaId) {
+    (get_from_seed::<AuraId>(s), get_from_seed::<GrandpaId>(s))
+}
+
+pub fn authority_keys_from_ss58(s_aura: &str, s_grandpa: &str) -> (AuraId, GrandpaId) {
+    (
+        get_aura_from_ss58_addr(s_aura),
+        get_grandpa_from_ss58_addr(s_grandpa),
+    )
+}
+
+pub fn get_aura_from_ss58_addr(s: &str) -> AuraId {
+    Ss58Codec::from_ss58check(s).unwrap()
+}
+
+pub fn get_grandpa_from_ss58_addr(s: &str) -> GrandpaId {
+    Ss58Codec::from_ss58check(s).unwrap()
+}
+
+// Includes for nakamoto genesis
+use serde::{Deserialize, Serialize};
+use serde_json as json;
+use std::{fs::File, path::PathBuf};
+
+// Configure storage from nakamoto data
+#[derive(Deserialize, Debug)]
+struct ColdkeyHotkeys {
+    stakes: std::collections::HashMap<String, std::collections::HashMap<String, (u64, u16)>>,
+    balances: std::collections::HashMap<String, u64>,
+}
